ran clean up on entire project

+++ src/main/java/net/telepathicgrunt/worldblender/blocks/WBPortalTileEntityRenderer.java
@@ -43,7 +43,7 @@ public class WBPortalTileEntityRenderer extends TileEntityRenderer<WBPortalTileE
 
 		for (int currentPass = 1; currentPass < passes; ++currentPass)
 		{
-			this.drawColor(tileEntity, 2.2F / (float) (20 - currentPass), matrix4f, renderBuffer.getBuffer(WB_RENDER_TYPE.get(currentPass)));
+			this.drawColor(tileEntity, 2.2F / (20 - currentPass), matrix4f, renderBuffer.getBuffer(WB_RENDER_TYPE.get(currentPass)));
 		}
 	}
